Claim voucher working temp
The PrizeView component now works to add a claim to the claim table based on the user and voucher selected, the voucher is also added to the user_vouchers table. it will show up accordingly on the vouchers' tab

Currently, no time limit is set. This will need to be set in the future. Mission weekly time reset and prize wheel daily time reset is needed